test: refactor exec-meshctl to use with meshadm
